vulpy/requirements.txt
click
cryptography
Flask
geoip2
jsonschema
PyJWT
pyotp
qrcode
requests
